Enhance fnancz2-installer1.sh for version 0.9.12
Updated download command to fetch multiple files and added error handling for the new fix file application.

+++ fnancz2-installer1.sh
@@ -92,17 +92,23 @@ chmod +x run.sh
 chmod +x game_updater.sh
 
 if [[ "$versionnumber" == "0.9.12" ]]; then
-   if wget https://github.com/AmrThePigeon/FNANCZAE2_Script_Builder/releases/download/0.9.12update/fnancz_2_yoyorunner_v0.9.12.fnancz; then
+   if wget https://github.com/AmrThePigeon/FNANCZAE2_Script_Builder/releases/download/0.9.12update/fnancz_2_yoyorunner_v0.9.12.fnancz && https://github.com/AmrThePigeon/FNANCZAE2_Script_Builder/releases/download/0.9.12update/fnancz2_update0.9.12_fix.fnanczupdate; then
       if xdelta3 -d -s Five_Nights_at_NCZ_2 fnancz_2_yoyorunner_v0.9.12.fnancz Five_Nights_at_NCZ_2_yoyo; then
-         rm fnancz_2_yoyorunner_v0.9.12.fnancz
-         rm Five_Nights_at_NCZ_2
-         mv Five_Nights_at_NCZ_2_yoyo Five_Nights_at_NCZ_2
-         echo -e "\033[0;34mCompiler switching for v0.9.12 complete\033[0m"
+         if xdelta3 -d -s "assets/game.unx" fnancz2_update0.9.12_fix.fnanczupdate "assets/game.unx"; then
+            rm fnancz2_update0.9.12_fix.fnanczupdate
          else
-           echo -e "\e[31mError while applying the new compiler, the game will won't run properly\e[0m"
-           rm fnancz_2_yoyorunner_v0.9.12.fnancz
-           rm Five_Nights_at_NCZ_2_yoyo
-           exit 1
+            rm fnancz2_update0.9.12_fix.fnanczupdate
+            echo -e "\e[31mError while applying the fix file, the game will won't run properly\e[0m"
+         fi
+      rm fnancz_2_yoyorunner_v0.9.12.fnancz
+      rm Five_Nights_at_NCZ_2
+      mv Five_Nights_at_NCZ_2_yoyo Five_Nights_at_NCZ_2
+      echo -e "\033[0;34mCompiler switching for v0.9.12 complete\033[0m"
+      else
+         echo -e "\e[31mError while applying the new compiler, the game will won't run properly\e[0m"
+         rm fnancz_2_yoyorunner_v0.9.12.fnancz
+         rm Five_Nights_at_NCZ_2_yoyo
+         exit 1
       fi
    else
      echo -e "\e[31mError: Failed to download assets, please check your internet connection then try again.\e[0m"
